[configuration] - simple parser scannerless not handling sub-expressions - to be continued

+++ core/app/terminal/src/main/java/com/dcall/core/app/terminal/gui/controller/keyboard/KeyboardAction.java
@@ -46,5 +46,5 @@ public enum KeyboardAction {
     public String getKeyControl() { return keyControl; }
     public Runnable getFunction() { return function; }
 
-    public int intValue() { return keyControl != null ? (int)keyControl.toCharArray()[0] : -1; }
+    public int intValue() { return keyControl != null ? keyControl.toCharArray()[0] : -1; }
 }
